+++ docs/source/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- score_only bool parameter to `moldrug.fitness.get_mol_cost`
+
 ## [2.1.7] - 2022.09.02
 
 ### Fixed